+++ lib/request_token.dart
@@ -15,28 +15,29 @@ class RequestToken {
 
   Future<Token> requestToken(String code) async {
     _generateTokenRequest(code);
-    return await _sendTokenRequest(_tokenRequest.url, _tokenRequest.params, _tokenRequest.headers);
+    return await _sendTokenRequest(
+        _tokenRequest.url, _tokenRequest.params, _tokenRequest.headers);
   }
 
-  Future<Token> requestRefreshToken(String refreshToken) async  {
+  Future<Token> requestRefreshToken(String refreshToken) async {
     _generateTokenRefreshRequest(refreshToken);
-    return await _sendTokenRequest(_tokenRefreshRequest.url, _tokenRefreshRequest.params, _tokenRefreshRequest.headers);
+    return await _sendTokenRequest(_tokenRefreshRequest.url,
+        _tokenRefreshRequest.params, _tokenRefreshRequest.headers);
   }
 
-  Future<Token> _sendTokenRequest(String url, Map<String, String> params, Map<String, String> headers) async {
-    Response response = await post(url,
-        body: params,
-        headers: headers);
+  Future<Token> _sendTokenRequest(String url, Map<String, String> params,
+      Map<String, String> headers) async {
+    var response = await post(url, body: params, headers: headers);
     Map<String, dynamic> tokenJson = json.decode(response.body);
-    Token token = new Token.fromJson(tokenJson);
+    var token = Token.fromJson(tokenJson);
     return token;
   }
 
   void _generateTokenRequest(String code) {
-    _tokenRequest = new TokenRequestDetails(config, code);
+    _tokenRequest = TokenRequestDetails(config, code);
   }
 
   void _generateTokenRefreshRequest(String refreshToken) {
-    _tokenRefreshRequest = new TokenRefreshRequestDetails(config, refreshToken);
-  }  
+    _tokenRefreshRequest = TokenRefreshRequestDetails(config, refreshToken);
+  }
 }