+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## [0.2.0] - 20201007
 
-- Add additional config options https://github.com/Earlybyte/aad_oauth/pull/66
-- Throw Exception on return https://github.com/Earlybyte/aad_oauth/pull/55
-- Fix Example App https://github.com/Earlybyte/aad_oauth/pull/65
+- Add additional config options #66
+- Throw Exception on return #55
+- Fix Example App #65
+- Comply Pub Dev Requirements #67
 
 ## [0.1.9] - 20200529